+++ client/app/src/main/java/edu/festival/deichkind/fragments/SettingsFragment.kt
@@ -1,11 +1,15 @@
 package edu.festival.deichkind.fragments
 
 import android.app.Activity
+import android.content.Context
 import android.os.Bundle
 import android.support.v7.preference.PreferenceFragmentCompat
 import edu.festival.deichkind.R
 import android.content.SharedPreferences
+import android.net.ConnectivityManager
+import android.support.v4.content.ContextCompat.getSystemService
 import android.support.v7.preference.ListPreference
+import android.widget.Toast
 import edu.festival.deichkind.SettingsActivity
 import edu.festival.deichkind.util.DykeManager
 import java.io.File
@@ -19,20 +23,26 @@ class SettingsFragment : PreferenceFragmentCompat(), SharedPreferences.OnSharedP
         preferenceManager.sharedPreferences.registerOnSharedPreferenceChangeListener(this)
 
         findPreference("sync-data-now").setOnPreferenceClickListener {
-            val dykesFile = File(context?.filesDir, "dykes.json")
-            val reportsFile = File(context?.filesDir, "reports.json")
+            val networkInfo = (context?.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager).activeNetworkInfo
 
-            if (dykesFile.exists()) {
-                dykesFile.delete()
-            }
+            if (networkInfo != null && networkInfo.isConnected) {
+                val dykesFile = File(context?.filesDir, "dykes.json")
+                val reportsFile = File(context?.filesDir, "reports.json")
 
-            if (reportsFile.exists()) {
-                reportsFile.delete()
-            }
+                if (dykesFile.exists()) {
+                    dykesFile.delete()
+                }
 
-            (activity as SettingsActivity).apply {
-                setResult(Activity.RESULT_OK)
-                finish()
+                if (reportsFile.exists()) {
+                    reportsFile.delete()
+                }
+
+                (activity as SettingsActivity).apply {
+                    setResult(Activity.RESULT_OK)
+                    finish()
+                }
+            } else {
+                Toast.makeText(context as Context, getString(R.string.main_network_connection_required_notice), Toast.LENGTH_LONG).show()
             }
 
             true
